requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.12.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
credentials==1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
emoji==2.12.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fasttext-wheel==0.9.2
filelock==3.14.0
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
geckodriver-autoinstaller==0.1.0
gensim==4.3.2
greenlet==3.0.3
h11==0.14.0
huggingface-hub==0.23.1
idna==3.4
install==1.3.5
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mongoengine==0.27.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.12.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.3.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
selenium==4.20.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_fastlang==2.1.0
speaklater==1.3
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
trio==0.25.0
trio-websocket==0.11.1
triton==2.3.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.0.1
XlsxWriter==3.2.0
yarl==1.9.4
